food_bot/requirements.txt
aiogram==3.*
aiohttp==3.*
python-dotenv==1.*
asyncio
matplotlib==3.*
